Train_Autolabel_Model/requirements.txt
pyarrow==17.0.0
torch
torchvision
evaluate
transformers
accelerate
datasets
Pillow
albumentations==1.2.1
huggingface_hub